% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onehot_encode}
\alias{onehot_encode}
\title{One-hot-encoding of the labels in case of classification}
\usage{
onehot_encode(y)
}
\arguments{
\item{y}{a numeric vector consisting of the response variable labels. The minimum value of the unique labels should begin from 0}
}
\description{
One-hot-encoding of the labels in case of classification
}
\examples{

library(elmNNRcpp)

y = sample(0:3, 100, replace = TRUE)

y_expand = onehot_encode(y)

}
