% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm.R
\name{predict.elm}
\alias{predict.elm}
\title{Predict with elm}
\usage{
\method{predict}{elm}(object, newdata, type = c("class", "prob", "raw"), ...)
}
\arguments{
\item{object}{elm model fitted with \code{\link{elm}}.}

\item{newdata}{data.frame with the new data}

\item{type}{only used with classification, can be either "class", "prob", "raw", 
which are class (vector), probability (matrix) or the output of the elm function (matrix).}

\item{...}{not used}
}
\value{
predicted values
}
\description{
Wrapper for \code{\link{elm_predict}}.
}
